# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QA Pulse - QA Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get user stats (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"

  # Projects
...
  /test-cases:
    get:
      operationId: listTestCases
      tags: [testCases]
      summary: List test cases
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: requirementId
          in: query
          schema:
            type: integer
        - name: authorId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
        - name: aiAssisted
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of test cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestCase"
    post:
      operationId: createTestCase
      tags: [testCases]
      summary: Create test case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestCaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
  /test-cases/{id}:
    get:
      operationId: getTestCase
      tags: [testCases]
      summary: Get test case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
    patch:
      operationId: updateTestCase
      tags: [testCases]
      summary: Update test case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestCaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"
    delete:
      operationId: deleteTestCase
      tags: [testCases]
      summary: Delete test case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /test-cases/{id}/clone:
    post:
      operationId: cloneTestCase
      tags: [testCases]
      summary: Clone test case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Cloned test case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestCase"

  # AI Test Case Generation
  /test-cases/ai-generate:
    post:
      operationId: generateTestCasesWithAI
      tags: [ai]
      summary: Generate test cases using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIGenerateInput"
      responses:
        "200":
          description: AI-generated test case suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIGenerateResponse"

  # Tasks
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          description: "qa_member | qa_lead | admin"
        team:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        mustChangePassword:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
        team:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        team:
          type: string
        avatarUrl:
          type: string
        mustChangePassword:
          type: boolean
        password:
          type: string
    UserStats:
      type: object
      required: [userId, tasksCompleted, tasksPending, tasksBlocked, tasksOverdue, testCasesCreated, bugsReported, bugsVerified, onTimeRate, overdueRate, automationContribution]
      properties:
        userId:
          type: integer
        tasksCompleted:
          type: integer
        tasksPending:
          type: integer
        tasksBlocked:
          type: integer
        tasksOverdue:
          type: integer
        testCasesCreated:
          type: integer
        bugsReported:
          type: integer
        bugsVerified:
          type: integer
        onTimeRate:
          type: number
        overdueRate:
          type: number
        automationContribution:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"

    # Project
...
    TestCase:
      type: object
      required: [id, title, type, priority, status, aiAssisted, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        objective:
          type: ["string", "null"]
        preconditions:
          type: ["string", "null"]
        testSteps:
          type: ["string", "null"]
        expectedResult:
          type: ["string", "null"]
        type:
          type: string
          description: "manual | automation_candidate"
        priority:
          type: string
        tags:
          type: ["string", "null"]
        requirementId:
          type: ["integer", "null"]
        requirementTitle:
          type: ["string", "null"]
        projectId:
          type: ["integer", "null"]
        projectName:
          type: ["string", "null"]
        linkedBug:
          type: ["string", "null"]
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        aiAssisted:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    TestCaseInput:
      type: object
      required: [title, type, priority]
      properties:
        title:
          type: string
        objective:
          type: string
        preconditions:
          type: string
        testSteps:
          type: string
        expectedResult:
          type: string
        type:
          type: string
        priority:
          type: string
        tags:
          type: string
        requirementId:
          type: integer
        projectId:
          type: integer
        linkedBug:
          type: string
        authorId:
          type: integer
        aiAssisted:
          type: boolean
        status:
          type: string
    TestCaseUpdate:
      type: object
      properties:
        title:
          type: string
        objective:
          type: string
        preconditions:
          type: string
        testSteps:
          type: string
        expectedResult:
          type: string
        type:
          type: string
        priority:
          type: string
        tags:
          type: string
        requirementId:
          type: integer
        linkedBug:
          type: string
        aiAssisted:
          type: boolean
        status:
          type: string

    # AI Generation
    AIGenerateInput:
      type: object
      required: [requirementTitle]
      properties:
        requirementTitle:
          type: string
        requirementDescription:
          type: string
        module:
          type: string
        projectId:
          type: integer
        release:
          type: string
        testCaseType:
          type: string
        priority:
          type: string
        tags:
          type: string
        additionalNotes:
          type: string
        requirementId:
          type: integer
        generatePositive:
          type: boolean
        generateNegative:
          type: boolean
        generateEdgeCases:
          type: boolean
        useSimilarHistorical:
          type: boolean
        useTemplateOnly:
          type: boolean
    AIGeneratedTestCase:
      type: object
      required: [title, objective, preconditions, testSteps, expectedResult, type, priority]
      properties:
        title:
          type: string
        objective:
          type: string
        preconditions:
          type: string
        testSteps:
          type: string
        expectedResult:
          type: string
        type:
          type: string
        priority:
          type: string
        tags:
          type: string
        automationCandidate:
          type: boolean
    AIGenerateResponse:
      type: object
      required: [testCases]
      properties:
        testCases:
          type: array
          items:
            $ref: "#/components/schemas/AIGeneratedTestCase"
        similarTestCasesUsed:
          type: integer
        templateUsed:
          type: string

    # Task
...
    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        entityId:
          type: ["integer", "null"]
        entityType:
          type: ["string", "null"]
        createdAt:
          type: string
...