# lib/api-zod/src/generated/types/aIGeneratedTestCase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QA Pulse - QA Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AIGeneratedTestCase {
  title: string;
  objective: string;
  preconditions: string;
  testSteps: string;
  expectedResult: string;
  type: string;
  priority: string;
  tags?: string;
  automationCandidate?: boolean;
}